pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hushhush"
version = "1.0.0"
description = "Password-protect your data. A modern, drop-in replacement for privy."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kalvad Tech" },
]
keywords = ["encryption", "password", "secret", "argon2", "fernet"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "argon2-cffi>=21.0",
    "cryptography>=3.0",
]

[project.urls]
Homepage = "https://github.com/KalvadTech/hushhush"
Repository = "https://github.com/KalvadTech/hushhush"

[tool.hatch.build.targets.wheel]
packages = ["src/hushhush"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
